Gameplay OptionIcon code cleanup
This commit cleans up the code related to displaying OptionIcons on ScreenGameplay.  Mostly, this involved condensing the code for P1 and P2 into a single file.  Previously, the code was effectively duplicated across two files.

I used the newer PlayerOptions interface where possible as it is more likely to exist in SM5.1 and beyond.  I couldn't find a way with PlayerOptions to determine whether "battery" was on or off so I stuck with the old-style check there.

Speed modifiers now use a BitmapText overlaid on top of a blank ("non.png") background.

It is my hope that this code will be easier to maintain.  Please ask if you have any questions.